requirements.txt
Django>=6.0.4
django-mptt>=0.14
Pillow>=10.0.0
psycopg2-binary>=2.9